[feat(l293d.c)] Implement l293d drivers
If included, commit implements the l293d motor driver user APIs that will be used throughout the project to control motor speed and direction. The driver includes APIs to set direction to forward, reverse, or stop values. Additionally includes an API that allows the user to set speed of the a given motor by specifying a desired pwm duty cycle value.